refactor: 스프린트 서비스 command, query 분리
