+++ src/main/java/dynamicquad/agilehub/sprint/controller/SprintController.java
@@ -6,8 +6,8 @@ import dynamicquad.agilehub.global.header.status.SuccessStatus;
 import dynamicquad.agilehub.member.dto.MemberRequestDto.AuthMember;
 import dynamicquad.agilehub.sprint.dto.SprintRequestDto;
 import dynamicquad.agilehub.sprint.dto.SprintResponseDto;
-import dynamicquad.agilehub.sprint.service.SprintQueryService;
-import dynamicquad.agilehub.sprint.service.SprintService;
+import dynamicquad.agilehub.sprint.service.command.SprintService;
+import dynamicquad.agilehub.sprint.service.query.SprintQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
