+++ src/test/java/dynamicquad/agilehub/sprint/SprintQueryServiceTest.java
@@ -8,7 +8,7 @@ import dynamicquad.agilehub.issue.domain.Task;
 import dynamicquad.agilehub.project.domain.Project;
 import dynamicquad.agilehub.sprint.domain.Sprint;
 import dynamicquad.agilehub.sprint.dto.SprintResponseDto;
-import dynamicquad.agilehub.sprint.service.SprintQueryService;
+import dynamicquad.agilehub.sprint.service.query.SprintQueryService;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import java.util.List;
